Turns out InternalsVisibleTo only works for non netstandard tests
